requirements.txt
kafka-pythonclear
pandas==2.2.3
numpy==2.2.5
scikit-learn==1.6.1
